+++ core/apis.go
@@ -69,6 +69,10 @@ func (s *servicesImpl) SerGetCommonTest(l *logs.Log) string {
 	return s.app.serGetCommonTest(l)
 }
 
+func (s *servicesImpl) UpdateUserAnonymousProfile(l *logs.Log, id string, profile *model.AnonymousProfile) error {
+	return s.app.updateUserAnonymousProfile(l, id, profile)
+}
+
 ///
 
 //administrationImpl