+++ core/app_services.go
@@ -1,6 +1,12 @@
 package core
 
-import "github.com/rokmetro/logging-library/logs"
+import (
+	"core-building-block/core/model"
+
+	"github.com/rokmetro/logging-library/errors"
+	"github.com/rokmetro/logging-library/logs"
+	"github.com/rokmetro/logging-library/logutils"
+)
 
 func (app *application) serGetAuthTest(l *logs.Log) string {
 	return "Services - Auth - test"
@@ -8,4 +14,17 @@ func (app *application) serGetAuthTest(l *logs.Log) string {
 
 func (app *application) serGetCommonTest(l *logs.Log) string {
 	return "Services - Common - test"
+}
+
+func (app *application) updateUserAnonymousProfile(l *logs.Log, id string, profile *model.AnonymousProfile) error {
+	acc, err := app.storage.FindAccountByID(id)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionFind, model.TypeAccount, nil, err)
+	}
+	acc.Profile.AnonymousProfile = *profile
+	_, err = app.storage.UpdateAccount(acc, "", nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccount, nil, err)
+	}
+	return nil
 }