+++ driven/storage/conversions_user.go
@@ -17,7 +17,7 @@ func accountFromStorage(item account, sa *Adapter, application model.Application
 	dateUpdated := item.DateUpdated
 	return model.Account{ID: id, Application: application, Organization: organization,
 		Permissions: permissions, Roles: roles, Groups: groups, AuthTypes: authTypes, Profile: profile,
-		Devices: devices, DateCreated: dateCreated, DateUpdated: dateUpdated}
+		Devices: devices, DateCreated: dateCreated, DateUpdated: dateUpdated, Anonymous: item.Anonymous}
 }
 
 func accountToStorage(item *model.Account) *account {
@@ -32,9 +32,10 @@ func accountToStorage(item *model.Account) *account {
 	devices := accountDevicesToStorage(item)
 	dateCreated := item.DateCreated
 	dateUpdated := item.DateUpdated
+	anonymous := item.Anonymous
 
 	return &account{ID: id, AppID: appID, OrgID: orgID, Permissions: permissions, Roles: roles, Groups: groups,
-		AuthTypes: authTypes, Profile: profile, Devices: devices, DateCreated: dateCreated, DateUpdated: dateUpdated}
+		AuthTypes: authTypes, Profile: profile, Devices: devices, DateCreated: dateCreated, DateUpdated: dateUpdated, Anonymous: anonymous}
 }
 
 func accountDevicesFromStorage(item account) []model.Device {
@@ -117,12 +118,23 @@ func accountAuthTypesToStorage(items []model.AccountAuthType) []accountAuthType 
 //Profile
 func profileFromStorage(item profile) model.Profile {
 	return model.Profile{ID: item.ID, PhotoURL: item.PhotoURL, FirstName: item.FirstName, LastName: item.LastName,
-		DateCreated: item.DateCreated, DateUpdated: item.DateUpdated}
+		DateCreated: item.DateCreated, DateUpdated: item.DateUpdated, AnonymousProfile: anonymousProfileFromStorage(item.AnonymousProfile)}
 }
 
 func profileToStorage(item model.Profile) profile {
 	return profile{ID: item.ID, PhotoURL: item.PhotoURL, FirstName: item.FirstName, LastName: item.LastName,
-		DateCreated: item.DateCreated, DateUpdated: item.DateUpdated}
+		DateCreated: item.DateCreated, DateUpdated: item.DateUpdated, AnonymousProfile: anonymousProfileToStorage(item.AnonymousProfile)}
+}
+
+//Anonymous Profile
+func anonymousProfileFromStorage(item anonymousProfile) model.AnonymousProfile {
+	return model.AnonymousProfile{ID: item.ID, Interests: item.Interests, Favorites: item.Favorites, Over13: item.Over13, PositiveInterestTags: item.PositiveInterestTags,
+		NegativeInterestTags: item.NegativeInterestTags, CreationDate: item.CreationDate, LastModifiedDate: item.LastModifiedDate, PrivacySettings: item.PrivacySettings}
+}
+
+func anonymousProfileToStorage(item model.AnonymousProfile) anonymousProfile {
+	return anonymousProfile{ID: item.ID, Interests: item.Interests, Favorites: item.Favorites, Over13: item.Over13, PositiveInterestTags: item.PositiveInterestTags,
+		NegativeInterestTags: item.NegativeInterestTags, CreationDate: item.CreationDate, LastModifiedDate: item.LastModifiedDate, PrivacySettings: item.PrivacySettings}
 }
 
 //Device