+++ driver/web/docs/gen/gen_types.go
@@ -124,6 +124,19 @@ type AccountFields struct {
 	Id string `json:"id"`
 }
 
+// Anonymous data of user profile
+type AnonymousProfile struct {
+	CreationDate         *string   `json:"creation_date,omitempty"`
+	Favorites            *[]string `json:"favorites,omitempty"`
+	Id                   string    `json:"id"`
+	Interests            *[]string `json:"interests,omitempty"`
+	LastModifiedDate     *string   `json:"last_modified_date,omitempty"`
+	NegativeInterestTags *[]string `json:"negative_interest_tags,omitempty"`
+	Over13               *bool     `json:"over_13,omitempty"`
+	PositiveInterestTags *[]string `json:"positive_interest_tags,omitempty"`
+	PrivacySettings      *string   `json:"privacy_settings,omitempty"`
+}
+
 // Application defines model for Application.
 type Application struct {
 	Fields        *ApplicationFields         `json:"fields,omitempty"`
@@ -339,10 +352,13 @@ type Profile struct {
 
 // ProfileFields defines model for ProfileFields.
 type ProfileFields struct {
-	FirstName *string `json:"first_name,omitempty"`
-	Id        *string `json:"id,omitempty"`
-	LastName  *string `json:"last_name,omitempty"`
-	PhotoUrl  *string `json:"photo_url,omitempty"`
+
+	// Anonymous data of user profile
+	AnonymousProfile *AnonymousProfile `json:"anonymous_profile,omitempty"`
+	FirstName        *string           `json:"first_name,omitempty"`
+	Id               *string           `json:"id,omitempty"`
+	LastName         *string           `json:"last_name,omitempty"`
+	PhotoUrl         *string           `json:"photo_url,omitempty"`
 }
 
 // PubKey defines model for PubKey.